Time: 13 ms (24.23%), Space: 156.3 MB (42.03%) - LeetHub

+++ 2140-solving-questions-with-brainpower/2140-solving-questions-with-brainpower.java
@@ -2,7 +2,7 @@ class Solution {
     long helper(int index, int[][] questions, long[] dp){
         if(index>=questions.length) return 0;
         if(dp[index] != -1L) return dp[index];
-        long point = questions[index][0];
+        int point = questions[index][0];
         int brainpower = questions[index][1]+1;
         long solve = point + helper(index+brainpower, questions, dp);
         long skip = helper(index+1, questions, dp);
